feat(providers): add descriptions and helper functions for Composio toolkit integrations
- Introduced a new `descriptions.rs` file containing human-readable capability summaries for various Composio toolkit slugs, enhancing the prompt renderer's ability to inform the orchestrator about connected integrations.
- Added a `helpers.rs` file with a utility function `pick_str` to streamline JSON object parsing for user profile extraction across different providers.
- Updated `mod.rs` to include the new modules, improving code organization and maintainability.
- Created a `traits.rs` file defining the core provider trait for Composio toolkit implementations, ensuring a consistent interface for all providers.
- Established a `types.rs` file to define shared types used across provider implementations, including sync reasons and user profile structures.
